requirements.txt
gpxpy
srtm.py
dateparser